backend/requirements.txt
altair==5.3.0
asgiref==3.8.1
attrs==23.2.0
blinker==1.8.2
cachetools==5.4.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0.7
djangorestframework==3.15.2
eurostatapiclient==0.3.0
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
Jinja2==3.1.4
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
numpy==2.0.0
packaging==24.1
pandas==2.2.2
pillow==10.4.0
prettytable==3.10.2
protobuf==5.27.2
pyarrow==17.0.0
pydeck==0.9.1
pyecharts==2.0.6
Pygments==2.18.0
python-dateutil==2.9.0.post0
pytz==2024.1
referencing==0.35.1
requests==2.32.3
rich==13.7.1
rpds-py==0.19.1
simplejson==3.19.2
six==1.16.0
smmap==5.0.1
sqlparse==0.5.1
streamlit==1.37.0
streamlit-echarts==0.4.0
tenacity==8.5.0
toml==0.10.2
toolz==0.12.1
tornado==6.4.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
watchdog==4.0.1
wcwidth==0.2.13
